requirements.txt
python-dotenv
litellm
rich
ollama
langchain-community
langchain_neo4j
rank_bm25
pydantic>=2.0.0
